+++ src/mocks/index.js
@@ -1,0 +1,6 @@
+export { authHandler } from "./auth";
+export { usersHandler } from "./users";
+export { boardHandler } from "./board";
+export { imageHandler } from "./image";
+export { allPostsData } from "./data/allPostsData";
+export { rabbitPostsData } from "./data/rabbitPostsData";
